Put the full path to the config.h file.
CMake 3.20 complains if you give it files it can't find.
We only see this on Windows right now since that is the
only platform with CMake 3.20.

It must have been the case that this file couldn't be found
all along, but we just never noticed it.  Fix it here by
giving the full path to where the file is generated.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>